streamlit_app/requirements.txt
db-dtypes
google-cloud-bigquery
pandas
plotly
streamlit
pygwalker
python-dotenv
dotenv